% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegSchaefer-package.R
\docType{package}
\name{ggsegSchaefer-package}
\alias{ggsegSchaefer}
\alias{ggsegSchaefer-package}
\title{ggsegSchaefer: Schaefer Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Schaefer cortical atlas for the 'ggseg' ecosystem. Provides unified 'ggseg_atlas' objects with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'. Includes 7-network and 17-network parcellations at resolutions from 100 to 1000 parcels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegSchaefer}
  \item Report bugs at \url{https://github.com/ggseg/ggsegSchaefer/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

Other contributors:
\itemize{
  \item Monica Thieu \email{monica.thieu@columbia.edu} (\href{https://orcid.org/0000-0002-6376-0467}{ORCID}) [contributor]
}

}
\keyword{internal}
