% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{schaefer17_200}
\alias{schaefer17_200}
\title{Schaefer 17-Network 200-Parcel Atlas}
\usage{
schaefer17_200()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Schaefer et al. (2018) local-global parcellation
of the human cerebral cortex with 200 parcels aligned to 17 resting-state
networks (Yeo et al., 2011).
Contains both 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
schaefer17_200()
}
\references{
Schaefer A, et al. (2018). "Local-Global Parcellation of the
Human Cerebral Cortex from Intrinsic Functional Connectivity MRI."
\emph{Cerebral Cortex}, 28(9):3095-3114.
\doi{10.1093/cercor/bhx179}
}
\seealso{
Other ggseg_atlases: 
\code{\link{schaefer17_100}()},
\code{\link{schaefer17_1000}()},
\code{\link{schaefer17_300}()},
\code{\link{schaefer17_400}()},
\code{\link{schaefer17_500}()},
\code{\link{schaefer17_600}()},
\code{\link{schaefer17_700}()},
\code{\link{schaefer17_800}()},
\code{\link{schaefer17_900}()},
\code{\link{schaefer7_100}()},
\code{\link{schaefer7_1000}()},
\code{\link{schaefer7_200}()},
\code{\link{schaefer7_300}()},
\code{\link{schaefer7_400}()},
\code{\link{schaefer7_500}()},
\code{\link{schaefer7_600}()},
\code{\link{schaefer7_700}()},
\code{\link{schaefer7_800}()},
\code{\link{schaefer7_900}()}
}
\concept{ggseg_atlases}
