% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\docType{data}
\name{schaefer7}
\alias{schaefer7}
\alias{schaefer7_100}
\alias{schaefer7_200}
\alias{schaefer7_300}
\alias{schaefer7_400}
\alias{schaefer7_500}
\alias{schaefer7_600}
\alias{schaefer7_700}
\alias{schaefer7_800}
\alias{schaefer7_900}
\alias{schaefer7_1000}
\title{Local-Global 7-Network Parcellation of the Human Cerebral Cortex}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.

An object of class \code{brain_atlas} of length 4.
}
\usage{
data(schaefer7_100)

schaefer7_200

schaefer7_300

schaefer7_400

schaefer7_500

schaefer7_600

schaefer7_700

schaefer7_800

schaefer7_900

schaefer7_1000
}
\description{
Coordinate data for the Schaefer et al. (2018) cortical atlas,
with 100-1000 regions on the cortical surface of the brain, aligned with
7 previously described functional networks (Yeo et al., 2011).
}
\examples{
data(schaefer7_100)
}
\references{
Schaefer et al. (2018) Cereb Cortex. 2018 Sep 1;28(9):3095-3114. doi: 10.1093/cercor/bhx179
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28981612}{PubMed})

\itemize{
 \item{schaefer7_100}{ 100 parellation of the 7 networks}
 \item{schaefer7_200}{ 200 parellation of the 7 networks}
 \item{schaefer7_300}{ 300 parellation of the 7 networks}
 \item{schaefer7_400}{ 400 parellation of the 7 networks}
 \item{schaefer7_500}{ 500 parellation of the 7 networks}
 \item{schaefer7_600}{ 600 parellation of the 7 networks}
 \item{schaefer7_700}{ 700 parellation of the 7 networks}
 \item{schaefer7_900}{ 800 parellation of the 7 networks}
 \item{schaefer7_900}{ 900 parellation of the 7 networks}
 \item{schaefer7_1000}{ 1000 parellation of th 7 networks}
}
}
\seealso{
Other ggseg_atlases: 
\code{\link{schaefer17}}
}
\concept{ggseg_atlases}
\keyword{datasets}
