% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\docType{data}
\name{schaefer7_3d}
\alias{schaefer7_3d}
\alias{schaefer7_400_3d}
\alias{schaefer7_100_3d}
\alias{schaefer7_200_3d}
\alias{schaefer7_300_3d}
\alias{schaefer7_500_3d}
\alias{schaefer7_600_3d}
\alias{schaefer7_700_3d}
\alias{schaefer7_800_3d}
\alias{schaefer7_900_3d}
\alias{schaefer7_1000_3d}
\title{Local-Global 7-Network Parcellation of the Human Cerebral Cortex}
\format{
A tibble with 4 observations and a nested data.frame
\describe{
  \item{surf}{type of surface (`inflated` or `white`)}
  \item{hemi}{hemisphere (`left`` or `right`)}
  \item{data}{data.frame of necessary variables for plotting
  }

  \item{atlas}{String. atlas name}
  \item{roi}{numbered region from surface}
  \item{annot}{concatenated region name}
  \item{label}{label `hemi_annot` of the region}
  \item{mesh}{list of meshes in two lists: vb and it}
  \item{area}{name of area in full}
  \item{colour}{HEX colour of region}
}

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.
}
\usage{
data(schaefer7_400_3d)

schaefer7_100_3d

schaefer7_200_3d

schaefer7_300_3d

schaefer7_500_3d

schaefer7_600_3d

schaefer7_700_3d

schaefer7_800_3d

schaefer7_900_3d

schaefer7_1000_3d
}
\description{
Mesh data for the 7 resting-state networks of
the Schaefer et al. (2018) networks.
}
\examples{
data(schaefer7_400_3d)
}
\references{
Schaefer et al. (2018) Cereb Cortex. 2018 Sep 1;28(9):3095-3114. doi: 10.1093/cercor/bhx179
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28981612}{PubMed})
}
\seealso{
Other ggseg3d_atlases: 
\code{\link{schaefer17_3d}}
}
\concept{ggseg3d_atlases}
\keyword{datasets}
