% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suit_atlas.R
\name{suit_atlas}
\alias{suit_atlas}
\title{Cerebellum SUIT atlas}
\usage{
suit_atlas()
}
\description{
Retrieve the cerebellar SUIT atlas provided by the ggsegExtra package.
}
\examples{
if(requireNamespace("ggsegExtra", quietly = TRUE)){
  ggseg::ggseg(atlas = suit_atlas())
}
}
