% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\docType{data}
\name{schaefer17}
\alias{schaefer17}
\title{Schaefer 7 network parcellation}
\format{
A data.frame with 10913 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{region}{name of region}
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral)}
  \item{label}{label from recon-all segmentation}
  \item{atlas}{name of the atlas}
}
}
\usage{
data(schaefer17)
}
\description{
Schaefer 7 network parcellation
}
\examples{
data(schaefer17)
}
\references{
Schaefer et al. (2018) Cereb Cortex. 2018 Sep 1;28(9):3095-3114. doi: 10.1093/cercor/bhx179
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28981612}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{schaefer17_200}},
\code{\link{schaefer17_400}},
\code{\link{schaefer7}}
}
\concept{ggseg_atlases}
\keyword{datasets}
