% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\docType{data}
\name{schaefer17}
\alias{schaefer17}
\alias{schaefer17_100}
\alias{schaefer17_200}
\title{Local-Global 17-Network Parcellation of the Human Cerebral Cortex}
\format{
A data.frame with 1 observation per ROI and 6 variables
\describe{
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral)}
  \item{region}{name of region}
  \item{label}{label from recon-all segmentation}
  \item{roi}{ROI number in the original atlas lookup table}
  \item{geometry}{\code{MULTIPOLYGON} of vertices for that region}
}

An object of class \code{brain_atlas} of length 4.
}
\usage{
data(schaefer17)

schaefer17_200
}
\description{
Coordinate data for the Schaefer et al. (2018) cortical atlas,
with 100-1000 regions on the cortical surface of the brain, aligned with
17 previously described functional networks (Yeo et al., 2011).
}
\examples{
data(schaefer17_100)
}
\references{
Schaefer et al. (2018) Cereb Cortex. 2018 Sep 1;28(9):3095-3114. doi: 10.1093/cercor/bhx179
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28981612}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{schaefer7}}
}
\concept{ggseg_atlases}
\keyword{datasets}
