% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\docType{data}
\name{schaefer7}
\alias{schaefer7}
\title{Local-Global 7 Parcellation of the Human Cerebral Cortex}
\format{
A data.frame with 10913 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{region}{name of region}
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral)}
  \item{acronym}{abbreviated name of region}
  \item{lobe}{lobe location}
  \item{label}{label from recon-all segmentation}
  \item{atlas}{name of the atlas}
}
}
\usage{
data(schaefer7)
}
\description{
Coordinate data for the Desikan-Killany Cortical atlas,
with 40 regions in on the cortical surface of the brain.
}
\examples{
data(schaefer7)
}
\references{
Schaefer et al. (2018) Cereb Cortex. 2018 Sep 1;28(9):3095-3114. doi: 10.1093/cercor/bhx179
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28981612}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{schaefer17}}
}
\concept{ggseg_atlases}
\keyword{datasets}
